requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
dnspython==2.7.0
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-ollama==0.3.5
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.8
langsmith==0.4.8
MarkupSafe==2.1.5
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.3
numpy==2.2.6
ollama==0.5.1
orjson==3.11.0
packaging==25.0
peft==0.16.0
pillow==11.0.0
portalocker==3.2.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pymongo==4.13.2
PyMuPDF==1.26.3
pymupdf4llm==0.0.27
pypdf==5.8.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
qdrant-client==1.15.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
sympy==1.13.3
tenacity==9.1.2
tokenizers==0.21.2
tqdm==4.67.1
transformers==4.53.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
zstandard==0.23.0
